Update POST method handler in product route to add a product to a cart
Create and execute queries to add a product to a cart. This involves updating the
quantity_available attribute of the product based on the quantity selected by the
customer and adding a new row to the products_carts table using the cart_id,
product_id, and product_quantity. Handle duplicate entries by replacing the old
record with the new record so that customers have the ability to change the quantity
of a product in their cart.