Fix error for orders with no customer ID and add comments
If an order was placed with a cart that had a NULL customer_id, the POST request to get the
customer_id would fail because of an attempt to index into a NoneType object. This was fixed
by removing the index.

Comments were added to the order route on server.py and to order.html.